+++ addon/components/citation-plugin/citations/search-modal.ts
@@ -1,4 +1,3 @@
-import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
 import { restartableTask, timeout } from 'ember-concurrency';
@@ -19,6 +18,8 @@ import {
   Option,
   unwrap,
 } from '@lblod/ember-rdfa-editor-lblod-plugins/utils/option';
+
+import PaginationComponent from '../../pagination/pagination-component';
 
 function getISODate(date: Option<Date>): string | null {
   if (date) {
@@ -43,14 +44,11 @@ interface Args {
   config: { endpoint: string };
 }
 
-export default class EditorPluginsCitationsSearchModalComponent extends Component<Args> {
+export default class EditorPluginsCitationsSearchModalComponent extends PaginationComponent<Args> {
   @service declare intl: IntlService;
   // Vlaamse Codex currently doesn't contain captions and content of decisions
   // @tracked isEnabledSearchCaption = false
   // @tracked isEnabledSearchContent = false
-  @tracked pageNumber = 0;
-  @tracked pageSize = 5;
-  @tracked totalCount = 0;
   @tracked decisions = [];
   @tracked error: unknown;
   @tracked selectedDecision: Decision | null = null;
@@ -106,23 +104,6 @@ export default class EditorPluginsCitationsSearchModalComponent extends Componen
 
   get searchText() {
     return this.inputSearchText ?? this.text;
-  }
-
-  get rangeStart() {
-    return this.pageNumber * this.pageSize + 1;
-  }
-
-  get rangeEnd() {
-    const end = this.rangeStart + this.pageSize - 1;
-    return end > this.totalCount ? this.totalCount : end;
-  }
-
-  get isFirstPage() {
-    return this.pageNumber == 0;
-  }
-
-  get isLastPage() {
-    return this.rangeEnd == this.totalCount;
   }
 
   resourceSearch = restartableTask(async () => {
@@ -223,18 +204,6 @@ export default class EditorPluginsCitationsSearchModalComponent extends Componen
   @action
   closeDecisionDetail() {
     this.selectedDecision = null;
-  }
-
-  // Pagination
-
-  @action
-  previousPage() {
-    --this.pageNumber;
-  }
-
-  @action
-  nextPage() {
-    ++this.pageNumber;
   }
 }