+++ addon/components/fragment-plugin/search-modal.ts
@@ -3,26 +3,29 @@ import { action } from '@ember/object';
 import { restartableTask, timeout } from 'ember-concurrency';
 import { task as trackedTask } from 'ember-resources/util/ember-concurrency';
 
-import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
 
 import { fetchFragments } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/fragment-plugin/utils/fetch-data';
 import { FragmentPluginConfig } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/fragment-plugin';
+
+import PaginationComponent from '../pagination/pagination-component';
 
 interface Args {
   config: FragmentPluginConfig;
   closeModal: () => void;
 }
 
-export default class FragmentPluginSearchModalComponent extends Component<Args> {
+export default class FragmentPluginSearchModalComponent extends PaginationComponent<Args> {
   // Filtering
   @tracked inputSearchText: string | null = null;
 
-  // Pagination
-  @tracked totalCount = 0;
-
   // Display
   @tracked error: unknown;
+
+  constructor(owner: unknown, args: Args) {
+    super(owner, args);
+    this.pageSize = 20;
+  }
 
   @action
   setInputSearchText(event: InputEvent) {
@@ -52,9 +55,14 @@ export default class FragmentPluginSearchModalComponent extends Component<Args> 
         filter: {
           name: this.inputSearchText ?? undefined,
         },
+        pagination: {
+          pageNumber: this.pageNumber,
+          pageSize: this.pageSize,
+        },
       });
 
       this.totalCount = queryResult.totalCount;
+
       return queryResult.results;
     } catch (error) {
       this.error = error;